+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Globals/GlobalDeclarations.cs
@@ -11,6 +11,7 @@
 
 global using System;
 global using System.Collections;
+global using System.Collections.Concurrent;
 global using System.Collections.Generic;
 global using System.Collections.ObjectModel;
 global using System.Collections.Specialized;
@@ -27,6 +28,7 @@ global using System.Drawing.Text;
 global using System.Globalization;
 global using System.IO;
 global using System.Linq;
+global using System.Management;
 #if NET8_0_OR_GREATER
 global using System.Text.Json.Nodes;
 #endif